refactor(lib): extract shared analyzeTokens module
